requirement.txt
pymongo
google-cloud-documentai
watchdog
PyMuPDF
pandas
fastapi
PYJWT
Pillow
bcrypt